pyproject.toml
[tool.poetry]
name = "django-twitter-authn-example"
version = "0.1.0"
description = ""
authors = ["kvwu <kvwu@transienterror.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
python-dotenv = "^0.18.0"
requests = "^2.25.1"
tweepy = {git = "https://github.com/tweepy/tweepy.git"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
